Add 'Quicklinks' (Download Address) as a Column
Added Quciklink (Downlaod Address) column to download list with the
option to show or hide the column Settings > Advanced > Download
Address Quicklink [Yes/No]

Allows user to "one-click" copy the url address to that download NOT
using shortcodes.

This is useful when placing a link to a download on an external
document or (non-wordpress) site.